+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# TAG v1.1.91 - Test
+
+- fix central connection handler
+
 # TAG v1.1.90 - Test
 
 - fix missing when case
@@ -286,17 +290,22 @@
 - Making the ble code more null safe
 
 # TAG v1.1.21 - Stable
+
 - Try to fix null error if device name is null
 
 # TAG v1.1.20 - Stable
+
 - Central.swift: Add removeAllServices() function call in stopAdvertising()
 
 # Tag v1.1.19 - Stable
+
 - Fixed changes from Tag 1.1.18
 
 # Tag v1.1.18
+
 - Changed BehaviorSubject to PublishSubject
 - Code has errors, don't use it
 
 # Tag v1.1.17 - Stable
+
 - Current working version